Liquids Theory

NESCGLE.ChebyshevGrids — Function

ChebyshevGrids(a :: Float64, b :: Float64, N :: Integer)

Returns a Chebyshev nodes and weights[1]

Arguments

a :: Float64: Initial node.

b :: Float64: Final node.

N :: Integer: Number of nodes.

References

[1] C.W. Clenshaw, A.R. Curtis, A method for numerical integration on an automatic computer, Numer. Math. 2 (1960) 197–205.

NESCGLE.UniformGrid — Function

UniformGrid(a :: Float64, b :: Float64, step :: Float64) = collect(a:step:b)

Returns a uniform grid. Is exactly the same as the fucntion collect.

Arguments

a :: Float64: Initial point.

b :: Float64: Final point.

step :: Float64: width step.

NESCGLE.ExponentialGrid — Function

ExponentialGrid(a :: Int64, b :: Int64, x₀ :: Float64)

Return an exponential grid of the form X = {x | x = x10^(n/a) for x in [0,b]} the sugested parameters are a = 5, b = 60.

Arguments

a :: Int64: Number por points between decade.

b :: Int64: Number of elements in the array.

x₀ :: Float64: Initial value.

NESCGLE.Cheb_plus_U — Function

Cheb_plus_U(x₀ :: Float64, x₁ :: Float64, x₂ :: Float64, N₁ :: Int64, N₂ :: Int64)

Auxiliar function that returns a composed grid by a Chebishev grid and a uniform grid.

Arguments

x₀ :: Float64: Initial point.

x₁ :: Float64: Middle point.

x₂ :: Float64: Final porint.

N₁ :: Int64: Number of point in the Chebishev grid.

N₂ :: Int64: number of point in the uniform grid.

NESCGLE.grid_plus_grid — Function

grid_plus_grid(x :: Array{Float64}, y :: Array{Float64})

Auxiliar fucntion to concat two grids.

Arguments

x :: Array{Float64}: First grid.

y :: Array{Float64}: Second grid.

NESCGLE.S_HS_PY — Function

S_HS_PY(ϕ :: Float64, k :: Float64) Returns the Static Structure Factor under the Percus-Yevick Closure[1].

\[S(k) = \frac{1}{1 - 24\phi c(k)}\]

Arguments

ϕ :: Float64: the volume fraction.

k :: Float64: wave vector.

References

[1] J. P. Hansen and I. McDonald. Theory of Simple Liquids. Academic, London, 1990.

S_HS_PY(ϕ::Float64)

Returns a function to construct the Static Structure Factor under the Percus-Yevick Closure[1].

\[S(k) = \frac{1}{1 - 24\phi c(k)}\]

Arguments

ϕ::Float64: the volume fraction.

References

[1] J. P. Hansen and I. McDonald. Theory of Simple Liquids. Academic, London, 1990.

NESCGLE.S_HS_VW — Function

S_HS_VW(ϕ::Float64)

Returns a function to compute the Static Structure Factor under the Percus-Yevick Closure[1] using the Verlet-Weis correction[2].

\[S(k) = \frac{1}{1 - 24\phi c(k)}\]

Arguments

ϕ::Float64: the volume fraction.

References

[1] J. P. Hansen and I. McDonald. Theory of Simple Liquids. Academic, London, 1990. [2] Loup Verlet and Jean-Jacques Weis. Phys. Rev. A 5, 939 – Published 1 February 1972

S_HS_VW(ϕ::Float64, k::Array{Float64})

Returns aan array that includes the Static Structure Factor under the Percus-Yevick Closure[1] using the Verlet-Weis correction[2].

\[S(k) = \frac{1}{1 - 24\phi c(k)}\]

Arguments

ϕ::Float64: the volume fraction.

References

[1] J. P. Hansen and I. McDonald. Theory of Simple Liquids. Academic, London, 1990. [2] Loup Verlet and Jean-Jacques Weis. Phys. Rev. A 5, 939 – Published 1 February 1972

NESCGLE.S_WCA_blip — Function

S_WCA_blip(ϕ::Float64, T::Float64; ν = 6)

Returns a function to compute the Static Structure Factor under the Percus-Yevick Closure[1] using the Verlet-Weis correction[2] and the blip function[3].

\[S(k) = \frac{1}{1 - 24\phi c(k)}\]

Arguments

ϕ::Float64: the volume fraction.

T::Real: Temperature.

Keywords

ν = 6: Parameter to module the softness of the potential.

References

[1] J. P. Hansen and I. McDonald. Theory of Simple Liquids. Academic, London, 1990. [2] Loup Verlet and Jean-Jacques Weis. Phys. Rev. A 5, 939 – Published 1 February 1972 [3] Luis Enrique Sánchez-Díaz, Pedro Ramírez-González, and Magdaleno Medina-Noyola Phys. Rev. E 87, 052306 – Published 22 May 2013

S_WCA_blip(ϕ::Float64, T::Float64; ν = 6)

Returns an array that includes the Static Structure Factor under the Percus-Yevick Closure[1] using the Verlet-Weis correction[2] and the blip function[3].

\[S(k) = \frac{1}{1 - 24\phi c(k)}\]

Arguments

ϕ::Float64: the volume fraction.

T::Real: Temperature.

k::Array{Float64}: wave vector.

Keywords

ν = 6: Parameter to module the softness of the potential.

References

[1] J. P. Hansen and I. McDonald. Theory of Simple Liquids. Academic, London, 1990. [2] Loup Verlet and Jean-Jacques Weis. Phys. Rev. A 5, 939 – Published 1 February 1972 [3] Luis Enrique Sánchez-Díaz, Pedro Ramírez-González, and Magdaleno Medina-Noyola Phys. Rev. E 87, 052306 – Published 22 May 2013

NESCGLE.βU — Function

βU(params::Array{Float64}, k::Float64, name::String)

Auxiliary function to return the Fourier Transform of the pair interaction potential to evaluate the Random Phase Approximation[1].

Arguments

params::Array{Float64}: List of parameters to evaluate the potential.

k::Float64: Wave vector.

name::String: Potential name selector.

References

[1] R.V. Sharma and K.C. Sharma. The structure factor and the transport properties of dense fluids having molecules with square well potential are possible generalizations. Physica A: Statistical Mechanics and its Applications, 89(1):213–218, 1977

NESCGLE.S_RPA — Function

S_RPA(params::Vector{Float64}, k::Float64; potential = "")

Returns the static structure factor using the random phase approximation[1] of the form

\[S(k) - \frac{1}{1 - 24\phi c(k) - \phi\beta u(k)}\]

where βu(k) is the Fourier transform of the pair interaction potential.

Arguments

params::Array{Float64}: List of parameters to evaluate the potential.

k::Float: wave vector.

Keywords

potential = ""::String: Potential name selector.

References

[1] R.V. Sharma and K.C. Sharma. The structure factor and the transport properties of dense fluids having molecules with square well potential are possible generalizations. Physica A: Statistical Mechanics and its Applications, 89(1):213–218, 1977

structure_factor(params::Vector{Float64}, potential = "")

Returns a function to compute the static structure factor using the random phase approximation[1] of the form

\[S(k) - \frac{1}{1 - 24\phi c(k) - \phi\beta u(k)}\]

where βu(k) is the Fourier transform of the pair interaction potential.

Arguments

params::Array{Float64}: List of parameters to evaluate the potential.

Keywords

potential = ""::String: Potential name selector.

References

[1] R.V. Sharma and K.C. Sharma. The structure factor and the transport properties of dense fluids having molecules with square well potential are possible generalizations. Physica A: Statistical Mechanics and its Applications, 89(1):213–218, 1977

S_RPA(params::Vector{Float64}, k::Vector{Float64}; potential = "")

Returns an array with the static structure factor using the random phase approximation[1] of the form

\[S(k) - \frac{1}{1 - 24\phi c(k) - \phi\beta u(k)}\]

where βu(k) is the Fourier transform of the pair interaction potential.

Arguments

params::Array{Float64}: List of parameters to evaluate the potential.

k::Vector{Float64}: Wave vector.

Keywords

potential = ""::String: Potential name selector.

References

[1] R.V. Sharma and K.C. Sharma. The structure factor and the transport properties of dense fluids having molecules with square well potential are possible generalizations. Physica A: Statistical Mechanics and its Applications, 89(1):213–218, 1977

NESCGLE.S_HS_Sticky — Function

S_HS_Sticky(τ::Float64, ϕ::Float64, k::Float64) Computes the static structure factor of a Stucky Hard Sphere.

Arguments

τ::Float64: Temperature.

ϕ::Float64: Volume fraction.

k::Float64: Wave vector.

References

[1]

Contributed by O. Joquín'Jaime

S_HS_Sticky(τ::Float64, ϕ::Float64, k::Float64)

Computes the static structure factor of a Stucky Hard Sphere.

Arguments

τ::Float64: Temperature.

ϕ::Float64: Volume fraction.

k::Float64: Wave vector.

References

[1]

Contributed by O. Joquín'Jaime

NESCGLE.g_HS — Function

g_HS(η, r)

Returns the pair distribution function for a Hard Spheres system. This is an analytoc solution to the Ornstein-Sernike equation under the Percus-Yevick closure[1].

Arguments

η::Real: volume fraction.

r::Real: distance.

References

[1] J. P. Hansen and I. McDonald. Theory of Simple Liquids. Academic, London, 1990.