Preparation Protocol

NESCGLE.PreparationProtocol — Type

PreparationProtocol(process::Vector{Any}, time::Vector{Float64}, str::String, rate::Float64)

Protocol of preparation for the colloidal system.

Fields

process::Vector{Any} represents the Array of parameters

time::Vector{Float64} represents the Array of time storage

str::String represents the Kind of process

rate::Float64 represents the Velocity of the process

NESCGLE.StaticProcess — Function

StaticProcess(params::Vector{Float64})

Prepararion protocol to use the Equilibrium version of the NESCGLE Theory.

NESCGLE.InstantaneousProcess — Function

InstantaneousProcess(paramsI::Vector{Float64}, paramsF::Vector{Float64})

Preparation protocol for a sudden change in thermodynamic conditions paramsI to parasmF.

NESCGLE.FiniteRate — Function

FiniteRate(paramsI::Vector{Float64}, paramsF::Vector{Float64}, α::Float64, N = 30)

Preparation protocol for a finite rate α change in thermodynamic conditions paramsI to parasmF. This protocol is an extension of the instantaneous process and is approximated by a sequence of N finite instantaneous stepes.

NESCGLE.Hysteresis — Function

Hysteresis(paramsI::Vector{Float64}, paramsF::Vector{Float64}, α::Float64, N = 30)

Preparation protocol for a hysteresis process at finite rate α between the thermodynamic conditions paramsI to parasmF. This protocol is an extension of the finite rate process and is approximated by a sequence of N finite instantaneous stepes for the initial evolution ramp and another N for the final evolutionr ramp.