Stability Matrix

Documentation for the StabilityMatrix

NESCGLE.StabilityMatrix — Type

StabilityMatrix

Input to evaluate the equation of the Self-Consistent Generalized Langevin Equation theory.

Fields

species::Integer: list of parameters to define the system of interest. The first parameter in the params array must be the volume fraction, and the Second must be the Temperature.

params::Array{Float64}: list of parameters to define the system of interest. The first parameter in the params array must be the volume fraction, and the Second must be the Temperature.

S::Function: Function to generate the static structure factor.

k::Array{Float64}: wave vector grid.

system::String: system identifier.

NESCGLE.wave_vector — Function

Returns the wave vector of an input

NESCGLE.structure_factor — Function

structure_factor(params::Vector{Float64}, system = "")

Returns a function to construct the static structure factor for a given system. The systems that are already programmed are:

"HardSphere": HS for PY[1].

"WCA": SS under the blip function approximation[3].

"SquareWell": Potential under the RPA.

"Yukawa": Potential under the RPA.

"SALR": Potential under the RPA.

"AsakuraOosawa1": Potential under the RPA.

"AsakuraOosawa2": Potential under the RPA.

"StickyHS": Sticky Hard Sphere.

Arguments

params::Vector{Float64}: list of parameters to evaluate the potential

Keywords

system = "": System of interest

Abbreviations

HS: Hars Sphere

PY: Percus-Yevick

SS: Soft Sphere

RPA: Random Phase Approximation[4]

References

[1] J. P. Hansen and I. McDonald. Theory of Simple Liquids. Academic, London, 1990. [2] Loup Verlet and Jean-Jacques Weis. Phys. Rev. A 5, 939 – Published 1 February 1972 [3] Luis Enrique Sánchez-Díaz, Pedro Ramírez-González, and Magdaleno Medina-Noyola Phys. Rev. E 87, 052306 – Published 22 May 2013 [4] R.V. Sharma and K.C. Sharma. The structure factor and the transport properties of dense fluids having molecules with square well potential are possible generalizations. Physica A: Statistical Mechanics and its Applications, 89(1):213–218, 1977

Returns an array with the Static Structure Factor from an input

NESCGLE.structure_factor_function — Function

Returns a function to construct the Static Structure Factor from an input

NESCGLE.number_of_species — Function

Returns the number of species of the system

NESCGLE.parameters — Function

Returns the parameters from an input

NESCGLE.volume_fraction — Function

Returns the volume fraction from an input

NESCGLE.system_str — Function

Returns the kind of system from an input

NESCGLE.SM_HS — Function

SM_HS(ϕ::Float64, k::Array{Float64}; VW = false::Bool)

Constructor of an input for a Hard Sphere system.

Arguments

ϕ::Float64: volume fraction.

k::Array{Float64}: wave vector array.

Keywords

VW = false::Bool: Verlet-Weis correction[1].

References

[1] Loup Verlet and Jean-Jacques Weis. Phys. Rev. A 5, 939 – Published 1 February 1972

NESCGLE.SM_WCA — Function

SM_WCA(ϕ::Float64, T::Float64, k::Array{Float64}; ν = 6::Int64)

Constructor for a soft sphere with pair interaction defined by the Weeks–Chandler–Andersen potential[1].

Arguments

ϕ::Float64: volume fraction.

T::Float64: Temperature.

k::Array{Float64}: wave vector array.

Keywords

ν = 6::Int64: Parameter to module the softness of the potential.

References

[1] Luis Enrique Sánchez-Díaz, Pedro Ramírez-González, and Magdaleno Medina-Noyola Phys. Rev. E 87, 052306 – Published 22 May 2013

NESCGLE.SM_SW — Function

SM_SW(ϕ::Float64, T::Float64, λ::Float64, k::Array{Float64})

Constructor for an input of hard spheres with a Square Well attractive potential.

Arguments

ϕ::Float64: volume fraction.

T::Float64: Dimensionless temperature.

λ::Float64: range of the well in terms of the diameter sigma.

k::Array{Float64}: wave vector array.

NESCGLE.SM_Yukawa — Function

SM_Yukawa(ϕ::Float64, ϵ::Float64, z::Float64, k::Array{Float64})

Constructor for an input of hard spheres with a Yukawa attractive potential[1].

Arguments

ϕ::Float64: volume fraction.

ϵ::Float64: Aplitude potential (effective charge).

z::Float64: Inverse screening length.

k::Float64: Wave vector array.

References

[1] Yukawa, H. (1935). "On the interaction of elementary particles". Proc. Phys.-Math. Soc. Jpn. 17: 48.

NESCGLE.SM_SALR — Function

SM_SALR(ϕ::Float64, ϵ₁::Float64, z₁::Float64, ϵ₂::Float64, z₂::Float64, k::Array{Float64})

Constructor for an input of hard spheres with Sort Attractions and Long Repultions(SALR).

Arguments

ϕ::Float64: volume fraction.

ϵ₁::Float64: Aplitude potential (effective charge) for the attractive part.

z₁::Float64: Inverse screening length for the attractive part.

ϵ₂::Float64: Aplitude potential (effective charge) for the repulsive part.

z₂::Float64: Inverse screening length for the repulsive part.

k::Float64: Wave vector array.

NESCGLE.SM_AO — Function

SM_AO(ϕ::Float64, ϕₚ⁽ᴿ⁾ ::Float64, ξ::Float64, k::Array{Float64})

Constructor for an input of a mixture of colloid-polimers in the monodisperse approximation [1].

Arguments

ϕ::Float64: Volume fraction of the colloid.

ϕₚ⁽ᴿ⁾::Float64: Volume fraction of the polimer.

ξ::Float64: Radius of gyration of the polimer.

k::Array{Float64}: wave vector array.

References

[1] H. N. W. Lekkerkerker et al 1992 EPL 20 559

NESCGLE.SM_StickyHS — Function

SM_StikyHS(ϕ::Float64, τ::Float64, k::Array{Float64})

Constructor of an input for a Hard Sphere system.

Arguments

ϕ::Float64: volume fraction.

τ::Float64: .

k::Array{Float64}: wave vector array.

References

[1]

Contributed by O. Joaquín-Jaime

NESCGLE.update_SM — Function

update_SM(sm::StabilityMatrix, params::Array{Float64}) -> StabilityMatrix

Update the stability matrix with new parameters.

Arguments

sm::StabilityMatrix: An instance of the StabilityMatrix type representing the current state of the system.

params::Array{Float64}: The new parameters to update in the stability matrix.

Returns

StabilityMatrix: A new instance of the StabilityMatrix with the updated parameters.

Example

sm = StabilityMatrix(
    2, 
    [0.3, 300.0, 1.0, 1.5], 
    ["Volume Fraction", "Temperature", "Param3", "Param4"], 
    x -> 1.0 / (1.0 + x^2), 
    range(0.1, stop=10.0, length=100), 
    "Colloidal Suspension"
)
new_params = [0.4, 320.0, 1.2, 1.6]
updated_sm = update_SM(sm, new_params)
println(updated_sm)

update_SM(sm::StabilityMatrix, S::Function) -> StabilityMatrix

Update the structure factor of the stability matrix.

Arguments

sm::StabilityMatrix: An instance of the StabilityMatrix type representing the current state of the system.

S::Function: The new structure factor function to update in the stability matrix.

Returns

StabilityMatrix: A new instance of the StabilityMatrix with the updated structure factor.

Example

sm = StabilityMatrix(
    2, 
    [0.3, 300.0, 1.0, 1.5], 
    ["Volume Fraction", "Temperature", "Param3", "Param4"], 
    x -> 1.0 / (1.0 + x^2), 
    range(0.1, stop=10.0, length=100), 
    "Colloidal Suspension"
)
new_S = x -> 1.0 / (1.0 + 2*x^2)
updated_sm = update_SM(sm, new_S)
println(updated_sm)

NESCGLE.Copy — Function

Copy(sm::StabilityMatrix) -> StabilityMatrix

Create a copy of the given stability matrix.

Arguments

sm::StabilityMatrix: An instance of the StabilityMatrix type representing the current state of the system.

Returns

StabilityMatrix: A new instance of the StabilityMatrix that is a copy of the given stability matrix.

Example

sm = StabilityMatrix(
    2, 
    [0.3, 300.0, 1.0, 1.5], 
    ["Volume Fraction", "Temperature", "Param3", "Param4"], 
    x -> 1.0 / (1.0 + x^2), 
    range(0.1, stop=10.0, length=100), 
    "Colloidal Suspension"
)

sm_copy = Copy(sm)
println(sm_copy)

NESCGLE.Lyapunov_Stability — Function

Lyapunov_Stability(smI::StabilityMatrix, smF::StabilityMatrix) -> Tuple{Float64, Float64}

Calculate the Lyapunov stability and its derivative for a system transitioning from an initial to a final state.

Arguments

smI::StabilityMatrix: The stability matrix representing the initial state of the system.

smF::StabilityMatrix: The stability matrix representing the final state of the system.

Returns

Tuple{Float64, Float64}: A tuple containing:

Λ::Float64: The Lyapunov stability.

dΛ::Float64: The derivative of the Lyapunov stability.

Example

smI = StabilityMatrix(
    2, 
    [0.3, 300.0, 1.0, 1.5], 
    ["Volume Fraction", "Temperature", "Param3", "Param4"], 
    x -> 1.0 / (1.0 + x^2), 
    range(0.1, stop=10.0, length=100), 
    "Colloidal Suspension"
)
smF = StabilityMatrix(
    2, 
    [0.4, 320.0, 1.2, 1.6], 
    ["Volume Fraction", "Temperature", "Param3", "Param4"], 
    x -> 1.0 / (1.0 + 2*x^2), 
    range(0.1, stop=10.0, length=100), 
    "Colloidal Suspension"
)

Λ, dΛ = Lyapunov_Stability(smI, smF)
println("Lyapunov stability: ", Λ)
println("Derivative of Lyapunov stability: ", dΛ)