Utils

NESCGLE.save_data — Function

save_data(nombre, formato; header = "", flag = true)

Simple function to save data, this must be in the form [x y z ...] if you include a coma a transpose is saved.

Example

x = collect(1:10)
y = x.^2
save_data("example.dat", [x y])

NESCGLE.make_directory — Function

make_directory(path::String) -> String

Create a directory if it does not already exist, and return the path with a trailing slash.

Arguments

path::String: The path of the directory to create.

Returns

String: The path of the directory, with a trailing slash.

Example

dir_path = make_directory("data/experiment1")
println("Directory path: ", dir_path)

NESCGLE.bisection — Function

bisection(condition::Function, A::Real, T::Real, tolerance::Real; flag = false) Function to perform a sucessive bisection between two interest points.

According to Zeno's paradox, Achilles can never catch up with the tortoise because he has an infinite number of finite catch-ups to make. This apparent paradox is used to argue that motion is impossible and is simply an illusion. When Achilles reaches the point where the tortoise started the race, the tortoise will have already moved on. The paradox concerns a race between the fleet-footed Achilles and a slow-moving tortoise. The two start moving at the same moment, but if the tortoise is initially given a head start and continues to move ahead, Achilles can run at any speed and will never catch up with it.

Arguments

condition::Function: Criteria to perform a step.

A::Real: Initial condition.

T::Real: Maximum position of the target.

tolerance::Real: Tolerance.

Keywords

flag = false: flag to print the internal computing of the procedure.

NESCGLE.saving_options — Type

saving_options(SF::Bool, TP::Bool, tw::Bool, folder_creation::Bool)

A type representing the options for saving different aspects of the NESCGLE computation results.

Fields

SF::Bool: Indicates whether to save the structure factor.

TP::Bool: Indicates whether to save the transport properties.

tw::Bool: Indicates whether to save the waiting times.

folder_creation::Bool: Indicates whether to create a folder for saving the results.

Example

options = saving_options(true, false, true, true)

NESCGLE.waiting_times — Function

waiting_times(n::Integer, N::Integer, t0::Float64) -> ExponentialGrid

Generate an exponential grid of waiting times for the NESCGLE computation.

The waiting time is the time after a nonequilibrium process is started, such as a sudden quench in temperature or a slow compression. It is the time after a colloidal system is thermodynamically stressed.

Arguments

n::Integer: The number of waiting times to generate.

N::Integer: The total number of time steps in the exponential grid.

t0::Float64: The initial waiting time.

Returns

ExponentialGrid: An exponential grid of waiting times.

Example

tw = waiting_times(10, 100, 1.0)